requirements.txt
inquirer==2.10.0
python-dotenv==0.21.0
requests==2.25.1
